Remove fist API Routes and Modifications to CLI
