+++ BdsManegerInfo.json
@@ -1,142 +1,135 @@
 {
-    "docs_base": "https://docs.bdsmaneger.com/docs/Bds Maneger core",
-    "docs": {
-        "url": "https://docs.bdsmaneger.com",
-        "main": "Bds Maneger core",
-        "rest_api": "docs/Bds Maneger core/REST API/"
+  "temp_host": {
+    "url": "http://hosts.bdsmaneger.com:3020"
+  },
+  "Fetchs": {
+    "php": "https://raw.githubusercontent.com/The-Bds-Maneger/Php_Static_Binary/main/binarys.json"
+  },
+  "IgnoreLog": {
+    "bedrock": [
+      {
+        "value": "Running AutoCompaction",
+        "regex": true
+      }
+    ],
+    "java": [],
+    "pocketmine": [],
+    "spigot": [],
+    "dragonfly": []
+  },
+  "StartedServer": {
+    "bedrock": {
+      "value": "Server started",
+      "regex": true
     },
-    "temp_host": {
-        "url": "http://hosts.bdsmaneger.com:3020"
+    "java": {
+      "value": null,
+      "regex": false
     },
-    "Fetchs": {
-        "php": "https://raw.githubusercontent.com/The-Bds-Maneger/Php_Static_Binary/main/binarys.json",
-        "servers": "https://raw.githubusercontent.com/The-Bds-Maneger/external_files/main/Server.json"
+    "pocketmine": {
+      "value": null,
+      "regex": false
     },
-    "IgnoreLog": {
-        "bedrock": [
-            {
-                "value": "Running AutoCompaction",
-                "regex": true
-            }
-        ],
-        "java": [],
-        "pocketmine": [],
-        "spigot": [],
-        "dragonfly": []
+    "spigot": {
+      "value": null,
+      "regex": false
     },
-    "StartedServer": {
-        "bedrock": {
-            "value": "Server started",
-            "regex": true
-        },
-        "java": {
-            "value": null,
-            "regex": false
-        },
-        "pocketmine": {
-            "value": null,
-            "regex": false
-        },
-        "spigot": {
-            "value": null,
-            "regex": false
-        },
-        "dragonfly": {
-            "value": null,
-            "regex": false
-        }
+    "dragonfly": {
+      "value": null,
+      "regex": false
+    }
+  },
+  "Servers": {
+    "bedrock": {
+      "stop": {
+        "value": "stop",
+        "type": "command"
+      },
+      "op": {
+        "value": "op \"{{}}\"",
+        "type": "command"
+      },
+      "deop": "op \"{{Player}}\"",
+      "ban": "tp \"{{Player}}\" ~ ~99999 ~",
+      "kick": "kick \"{{Player}}\" \"{{Text}}\"",
+      "tp": "tp \"{{Player}}\" {{X}} {{Y}} {{X}}",
+      "say": "say {{Text}}"
     },
-    "Servers": {
-        "bedrock": {
-            "stop": {
-                "value": "stop",
-                "type": "command"
-            },
-            "op": {
-                "value": "op \"{{}}\"",
-                "type": "command"
-            },
-            "deop": "op \"{{Player}}\"",
-            "ban": "tp \"{{Player}}\" ~ ~99999 ~",
-            "kick": "kick \"{{Player}}\" \"{{Text}}\"",
-            "tp": "tp \"{{Player}}\" {{X}} {{Y}} {{X}}",
-            "say": "say {{Text}}"
-        },
-        "java": {
-            "stop": "stop",
-            "op": "op {{Player}}",
-            "deop": "op {{Player}}",
-            "ban": "ban {{Player}}",
-            "kick": "kick {{Player}} {{Text}}",
-            "tp": "tp {{Player}} {{X}} {{Y}} {{X}}",
-            "say": "say {{Text}}"
-        },
-        "pocketmine": {
-            "stop": {
-                "value": "stop",
-                "type": "command"
-            },
-            "op": {
-                "value": "op {{Player}}",
-                "type": "command"
-            },
-            "deop": {
-                "value": "deop {{Player}}",
-                "type": "command"
-            },
-            "ban": {
-                "value": "ban {{Player}}",
-                "type": "command"
-            },
-            "kick": {
-                "value": "kick {{Player}}",
-                "type": "command"
-            },
-            "tp": {
-                "value": "tp {{Player}} {{X}} {{Y}} {{X}}",
-                "type": "command"
-            },
-            "say": {
-                "value": "say {{Text}}",
-                "type": "command"
-            }
-        },
-        "dragonfly": {
-            "stop": {
-                "value": "SIGNKILL",
-                "type": "process"
-            },
-            "op": {
-                "value": null,
-                "type": null
-            },
-            "deop": {
-                "value": null,
-                "type": null
-            },
-            "ban": {
-                "value": null,
-                "type": null
-            },
-            "kick": {
-                "value": null,
-                "type": null
-            },
-            "tp": {
-                "value": null,
-                "type": null
-            },
-            "say": {
-                "value": null,
-                "type": null
-            }
-        }
+    "java": {
+      "stop": "stop",
+      "op": "op {{Player}}",
+      "deop": "op {{Player}}",
+      "ban": "ban {{Player}}",
+      "kick": "kick {{Player}} {{Text}}",
+      "tp": "tp {{Player}} {{X}} {{Y}} {{X}}",
+      "say": "say {{Text}}"
     },
-    "contributors": [
-        {
-            "name": "Matheus Sampaio Queiroga",
-            "email": "srherobrine20@gmail.com",
-            "url": "https://sirherobrine23.org/?from=the-bds-maneger-project"
-        }
-    ]
+    "pocketmine": {
+      "stop": {
+        "value": "stop",
+        "type": "command"
+      },
+      "op": {
+        "value": "op {{Player}}",
+        "type": "command"
+      },
+      "deop": {
+        "value": "deop {{Player}}",
+        "type": "command"
+      },
+      "ban": {
+        "value": "ban {{Player}}",
+        "type": "command"
+      },
+      "kick": {
+        "value": "kick {{Player}}",
+        "type": "command"
+      },
+      "tp": {
+        "value": "tp {{Player}} {{X}} {{Y}} {{X}}",
+        "type": "command"
+      },
+      "say": {
+        "value": "say {{Text}}",
+        "type": "command"
+      }
+    },
+    "dragonfly": {
+      "stop": {
+        "value": "SIGNKILL",
+        "type": "process"
+      },
+      "op": {
+        "value": null,
+        "type": null
+      },
+      "deop": {
+        "value": null,
+        "type": null
+      },
+      "ban": {
+        "value": null,
+        "type": null
+      },
+      "kick": {
+        "value": null,
+        "type": null
+      },
+      "tp": {
+        "value": null,
+        "type": null
+      },
+      "say": {
+        "value": null,
+        "type": null
+      }
+    }
+  },
+  "contributors": [
+    {
+      "name": "Matheus Sampaio Queiroga",
+      "email": "srherobrine20@gmail.com",
+      "url": "https://sirherobrine23.org/?from=the-bds-maneger-project"
+    }
+  ]
 }
