+++ bin/BdsManeger.js
@@ -1,6 +1,9 @@
 #!/usr/bin/env node
 if (process.platform === "win32") process.title = "Bds Maneger CLI"; else process.title = "Bds-Manger-CLI";
 process.env.IS_BDS_CLI = process.env.IS_BIN_BDS = true;
+// Internal Modules
+const fs = require("fs");
+const path = require("path");
 
 // External Modules
 const cli_color = require("cli-color");
@@ -14,9 +17,36 @@ const BdsCore = require("../index");
 const BdsNetwork = require("../src/BdsNetwork");
 const commandExits = require("../lib/commandExist");
 const readline = require("readline");
-// const BdsMenus = require("./bds_maneger/menus");
 const { PlatformVersionsV2 } = require("../src/BdsServersDownload");
 const { GetPlatform } = require("../lib/BdsSettings");
+
+// Load Bds Maneger CLI Plugins
+const MoreHelp = [];
+const BeforeRun = [];
+fs.readdirSync(path.join(__dirname, "plugins")).map(file => path.resolve(__dirname, "plugins", file)).filter(Mod => fs.lstatSync(Mod).isFile() && Mod.endsWith(".js")).forEach(Plugin => {
+  try {
+    const __module = require(Plugin);
+    (__module.Args || []).forEach(PluginArg => {
+      ["h", "help", "i", "info", "d", "download", "s", "start", "k", "kill", "get_domain", "p", "platform", "n", "no-api"].forEach(Arg => {
+        if (PluginArg.arg === Arg) {
+          console.log(cli_color.redBright(`${path.basename(Plugin).replace(/\.js$/gi, "")}:`, "Conflicted with Bds Maneger CLI argument"));
+          process.exit(12);
+        }
+      });
+      BeforeRun.forEach(Arg => {
+        if (PluginArg.arg === Arg) {
+          console.log(cli_color.redBright(`${path.basename(Plugin).replace(/\.js$/gi, "")}:`, "Conflicted with another plugin argument"));
+          process.exit(13);
+        }
+      });
+      BeforeRun.push(PluginArg);
+      MoreHelp.push(cli_color.redBright(`${path.basename(Plugin).replace(/\.js$/gi, "")} - ${__module.description}`), "", ...(__module.help || []), "");
+    });
+  } catch (err) {
+    console.log(cli_color.redBright(`Error loading plugin: ${Plugin}`));
+    console.log(cli_color.redBright(err));
+  }
+});
 
 async function DownloadServer() {
   const ora = (await import("ora")).default;
@@ -64,27 +94,33 @@ async function help() {
   const help = [
     `Bds Maneger CLI version: ${cli_color.magentaBright(BdsCore.package_json.version)}`,
     `System: ${cli_color.yellow(process.platform)}, architecture: ${cli_color.blue(BdsCore.arch)}`,
-    "**************************************************************",
-    "* Usage: bds-maneger-cli [options]",
-    "*",
-    "* Options:",
-    "*   -h, --help                 Print this help",
-    "*   -i, --info                 Print info about Bds Maneger Core and Platforms",
-    "*   -d, --download             Download a server",
-    "*   -s, --start                Start a server",
-    "*       --get_domain           Get temporary public domain to connect in to server or API",
-    "*   -p, --platform             Change the platform",
-    "*   -n, --no-api               Don't start the Bds Maneger API Rest",
-    "*",
-    "* Examples:",
-    "*   bds-maneger-cli -d",
-    "*   bds-maneger-cli -s",
-    "*   bds-maneger-cli -p bedrock",
-    "*   bds-maneger-cli -i",
-    "*   bds-maneger-cli -h",
-    "**************************************************************"
+    "",
+    " Usage: bds-maneger-cli [options]",
+    "",
+    " Options:",
+    "   -h, --help                 Print this help",
+    "   -i, --info                 Print info about Bds Maneger Core and Platforms",
+    "   -d, --download             Download a server",
+    "   -s, --start                Start a server",
+    "   -k, --kill                 Kill Server if running",
+    "       --get_domain           Get temporary public domain to connect in to server or API",
+    "   -p, --platform             Change the platform",
+    "   -n, --no-api               Don't start the Bds Maneger API Rest",
+    "",
+    ...MoreHelp,
+    "",
+    " Examples:",
+    "   bds-maneger-cli -d",
+    "   bds-maneger-cli -s",
+    "   bds-maneger-cli -sk",
+    "   bds-maneger-cli -k",
+    "   bds-maneger-cli -p bedrock",
+    "   bds-maneger-cli -i",
+    "   bds-maneger-cli -h",
+    ""
   ];
-  return console.log(cli_color.whiteBright(help.join("\n").replace(/true/gi, cli_color.greenBright("true")).replace(/false/gi, cli_color.redBright("false")).replace(/undefined/gi, cli_color.red("undefined"))));
+  console.log(cli_color.whiteBright(help.join("\n").replace(/true/gi, cli_color.greenBright("true")).replace(/false/gi, cli_color.redBright("false")).replace(/undefined/gi, cli_color.red("undefined"))));
+  return process.exit(0);
 }
 
 // Async functiona
@@ -136,13 +172,19 @@ async function Runner() {
     }
   }
 
-  const BdsCoreStart = BdsCore.start();
-  BdsCoreStart.log(data => console.log(cli_color.blueBright(data.replace(/\n$/gi, ""))));
-  BdsCoreStart.exit(code => {
-      console.log(cli_color.redBright(`Bds Core Exit with code ${code}, Uptimed: ${BdsCoreStart.uptime}`));
+  // Load Plugins
+  for (let Plugin of BeforeRun) {
+    if (!(ProcessArgs[Plugin.arg])) Plugin.main(ProcessArgs[Plugin.arg], ProcessArgs).catch(err => console.log("Plugin Crash:", "\n\n", err));
+  }
+
+  const BdsManegerServer = BdsCore.start();
+  BdsManegerServer.log(data => console.log(cli_color.blueBright(data.replace(/\n$/gi, ""))));
+  BdsManegerServer.exit(code => {
+      console.log(cli_color.redBright(`Bds Core Exit with code ${code}, Uptimed: ${BdsManegerServer.uptime}`));
       process.exit(code);
   });
   if (!(ProcessArgs["no-api"])) BdsCore.api();
-  readline
+  const __readline = readline.createInterface({input: process.stdin, output: process.stdout});
+  __readline.on("line", data => BdsManegerServer.command(data));
 }
 Runner();
