+++ lib/Requests.js
@@ -1,8 +1,7 @@
-// Load Fetch.js
-const fetchS = (...args) => import("node-fetch").then(mod => mod.default(...args));
-
-// Set global fetch
-if (typeof fetch === "undefined") global.fetch = fetchS;
+if (typeof fetch === "undefined") {
+    global.fetch = (...args) => import("node-fetch").then(mod => mod.default(...args));
+    import("node-fetch").then(mod => global.fetch = mod.default);
+}
 
 // Request Json
 const ReqJson = async (url = "", options) => await (await fetch(url, options)).json()
