+++ package.json
@@ -3,7 +3,7 @@
   "publishConfig": {
     "access": "public"
   },
-  "version": "1.14.1",
+  "version": "1.15.0",
   "description": "Scripts to manage minecraft server's",
   "private": false,
   "main": "index.js",
@@ -20,8 +20,7 @@
     "remove_dev_versions": "node .Build/RemoveVersions.js"
   },
   "bin": {
-    "bds_maneger": "bin/bds_maneger.js",
-    "bds_telegram": "bin/telegram_bot.js"
+    "bds_maneger": "bin/BdsManeger.js"
   },
   "repository": {
     "type": "git",
